clock: at91: Restore generic clocks on resume
The Peripheral Control Register in the PMC controller is used as a
window to access the configuration for 62 clock registers. To write
the configuration for a specific clock, the CMD bit must be set to 1,
otherwise the write only selects the register available for the next
read command.

Fix the resume code to follow this rule.