ARM: dts: at91: sama5d2: fix m_can Message RAM config
From the commit (44c5584e1e), The TXE FIFO is needed to be configured,
otherwise, the driver doesn't detect the transmission of the message.
So configures 32 TX Event FIFO elements and 32 TX Buffer elements and
change the Rx Buffer elements to 64 maximum as well, and configure the
offset of CAN1 at the end of CAN0's.

Signed-off-by: Wenyou Yang <wenyou.yang@atmel.com>