pinctrl: at91-pio4: add support for drive-strength property
Add support for the drive-strength property. Usually its value is
expressed in mA. Since the numeric value depends on VDDIOP voltage,
the controller uses low, medium and high to define the drive-strengh.

The pio controller accepts two values for the low drive: 0 or 1. As most
of the time it is not needed to change the drive, 0 is considered as the
default value. The low-drive value won't be advertised through
pinconf-pins file excepted if it has been set explicitly in the device
tree.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>