clk: at91: pmc: don't use timeout when resuming
When resuming the PMC, a call to regmap_read_poll_timeout() is used to
check if the UPLL is locked. Unfortunately, the timekeeping subsystem
may not be resumed yet.
Most of the time, the UPLL should be locked when resuming the PMC. If
it's not the case, print a warning. We don't want to stay in an infinite
loop if only peripherals using the UTMI clock won't work. Moreover, it
doesn't necessary lead to an issue. If the UPLL is locked before
peripherals use it, it won't be an issue.
If some unexpected behaviors are encountered for peripherals using the
UTMI clock, then the PMC_SR register should be checked manually to see
if the UPLL is locked or not.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>