input: atmel_ptc: add tracking support
The PTC sends too many IRQs, it causes to miss some of them. Most of the
time, it's not an issue since it concerns position reporting.
Unfortunately, we can miss IRQ about touch event. In this case, we can't
report that the scroller is no longer touched.

To avoid this situation, use a tasklet to perform a kind of tracking.
While tracking the position, IRQs received are discarded. To perform the
tracking, a tasklet is used to monitor the scroller position and status
until the status is set to "not touch".

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>