ARM: dts: at91: at91-onsite_te5000: BG96 power and reset lines
Due to an assortment of missing pull-ups and pull-downs, it's impossible
to perform a BG96 start-up that matches the timing diagrams in the
datasheet.

The CELL_PWR and CELL_RESET lines boot up with dueling pull-up/pull-down
pairs that leave them in an analog state that is high enough to saturate
the BJTs controlling PWRKEY and N_RESET, driving them both to 0 on
power-on when they are expected to be 1. Additionally, the FET driven by
3V8PWR is missing a pull-up so power to the BG96 can't be removed and
restored under software control.

Through experimentation I established that as long as you leave the
PWRKEY line low (CELL_PWR high) indefinitely, a rising edge on N_RESET
(falling edge on CELL_RESET) will reliably power on the BG96.

Without this careful arrangement of initial values, we experienced
intermittent failure to boot the BG96 at power-on on some boards.